⚙️ Move Library to features folder
