+++ client/src/AppRoutes.jsx
@@ -1,7 +1,7 @@
 import styles from './AppRoutes.module.scss';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Nav from './layout/Nav/Nav.jsx';
-import Library from './layout/Library/Library.jsx';
+import Library from './features/library/Library.jsx';
 
 const AppRoutes = () => {
   return (
