+++ client/src/features/library/LibraryHeader/LibraryFilter/LibraryFilter.tsx
@@ -2,7 +2,7 @@ import Tabs from '../../../../ui-library/Tabs/Tabs.jsx';
 import TabsList from '../../../../ui-library/Tabs/TabsList.jsx';
 import Tab from '../../../../ui-library/Tabs/Tab.jsx';
 import { useAppDispatch } from '../../../../app/hooks.ts';
-import { filterLibraryItems } from '../../../../features/library/librarySlice.ts';
+import { filterLibraryItems } from '../../librarySlice.ts';
 
 const LibraryFilter = () => {
   const dispatch = useAppDispatch();
