+++ client/src/features/library/LibraryList/LibraryCard/CardImage.tsx
