+++ client/src/features/library/LibraryList/LibraryCard/LibraryCard.module.scss
